+++ app/src/data/ffmpeg/ffmpegUtils.ts
@@ -1,4 +1,6 @@
 import { FFmpegSession } from 'ffmpeg-kit-react-native';
+import { Paths } from 'expo-file-system';
+import * as FileSystem from 'expo-file-system/legacy';
 
 /**
  * ユニークなファイル名サフィックスを生成する。
@@ -35,4 +37,28 @@ export function extractDurationFromLogs(logs: string): number | null {
   const m = parseInt(match[2], 10);
   const s = parseFloat(match[3]);
   return h * 3600 + m * 60 + s;
+}
+
+/**
+ * キャッシュディレクトリ内の古い一時出力ファイルを削除する。
+ * 対象: `_compressed_`, `_gabigabi_`, `_converted_`, `_passlog` を含むファイル名。
+ * (#177) FfmpegCompressor/Converter/Processor の重複実装を統合。
+ * (#183) passlogファイルもクリーンアップ対象に追加。
+ */
+export async function cleanupCachedTempFiles(): Promise<void> {
+  try {
+    const cacheDirUri = Paths.cache.uri;
+    const cacheDir = cacheDirUri.endsWith('/') ? cacheDirUri : cacheDirUri + '/';
+    const dirInfo = await FileSystem.getInfoAsync(cacheDir);
+    if (!dirInfo.exists) return;
+    const result = await FileSystem.readDirectoryAsync(cacheDir);
+    const tempPattern = /_(compressed|gabigabi|converted)_|_passlog/;
+    await Promise.all(
+      result
+        .filter(name => tempPattern.test(name))
+        .map(name => FileSystem.deleteAsync(cacheDir + name, { idempotent: true })),
+    );
+  } catch {
+    // クリーンアップ失敗は無視して処理を続行する
+  }
 }
